CovidTable: create file structure
