+++ site/client/covid/index.tsx
@@ -1,0 +1,11 @@
+import * as React from "react"
+import * as ReactDOM from "react-dom"
+
+import { CovidTable } from "./CovidTable"
+
+export function runCovid() {
+    const element = document.getElementById("covid-table-embed")
+    if (element) {
+        ReactDOM.render(<CovidTable />, element)
+    }
+}
